requirements.txt
flask
pandas
matplotlib
plotly
pywebview
pyinstaller
openpyxl
fpdf
